Move entity_id.h to public headers, format
