+++ src/apps/js_generic/js_generic.cpp
@@ -1526,9 +1526,7 @@ namespace ccfapp
       });
     }
 
-    void tick(
-      std::chrono::milliseconds elapsed,
-      size_t tx_count) override
+    void tick(std::chrono::milliseconds elapsed, size_t tx_count) override
     {
       metrics_tracker.tick(elapsed, tx_count);
