+++ src/consensus/aft/revealed_nonces.h
@@ -2,9 +2,9 @@
 // Licensed under the Apache 2.0 License.
 #pragma once
 
+#include "ccf/entity_id.h"
 #include "crypto/hash.h"
 #include "kv/map.h"
-#include "node/entity_id.h"
 
 #include <msgpack/msgpack.hpp>
 #include <string>
