+++ src/enclave/rpc_context.h
@@ -2,13 +2,13 @@
 // Licensed under the Apache 2.0 License.
 #pragma once
 
+#include "ccf/tx_id.h"
 #include "http/http_builder.h"
 #include "http/http_consts.h"
 #include "http/ws_consts.h"
 #include "node/client_signatures.h"
 #include "node/entities.h"
 #include "node/rpc/error.h"
-#include "ccf/tx_id.h"
 
 #include <llhttp/llhttp.h>
 #include <variant>
