+++ src/node/historical_queries.h
@@ -2,10 +2,10 @@
 // Licensed under the Apache 2.0 License.
 #pragma once
 
+#include "ccf/historical_queries_interface.h"
 #include "consensus/ledger_enclave_types.h"
 #include "ds/ccf_assert.h"
 #include "ds/spin_lock.h"
-#include "ccf/historical_queries_interface.h"
 #include "kv/store.h"
 #include "node/encryptor.h"
 #include "node/history.h"
@@ -251,10 +251,7 @@ namespace ccf::historical
 
                   auto proof = tree.get_proof(seqno);
                   details->receipt = std::make_shared<TxReceipt>(
-                    sig->sig,
-                    proof.get_root(),
-                    proof.get_path(),
-                    sig->node);
+                    sig->sig, proof.get_root(), proof.get_path(), sig->node);
                   details->transaction_id = {sig->view, seqno};
                   details->current_stage = RequestStage::Trusted;
                 }
@@ -288,10 +285,7 @@ namespace ccf::historical
 
                   auto proof = tree.get_proof(new_seqno);
                   details->receipt = std::make_shared<TxReceipt>(
-                    sig->sig,
-                    proof.get_root(),
-                    proof.get_path(),
-                    sig->node);
+                    sig->sig, proof.get_root(), proof.get_path(), sig->node);
                   details->transaction_id = {sig->view, new_seqno};
                   new_details->current_stage = RequestStage::Trusted;
                 }
@@ -319,12 +313,9 @@ namespace ccf::historical
                   return UpdateTrustedResult::Invalidated;
                 }
 
-                  auto proof = tree.get_proof(new_seqno);
-                  details->receipt = std::make_shared<TxReceipt>(
-                    sig->sig,
-                    proof.get_root(),
-                    proof.get_path(),
-                    sig->node);
+                auto proof = tree.get_proof(new_seqno);
+                details->receipt = std::make_shared<TxReceipt>(
+                  sig->sig, proof.get_root(), proof.get_path(), sig->node);
                 details->transaction_id = {sig->view, new_seqno};
                 new_details->current_stage = RequestStage::Trusted;
               }
