+++ src/node/members.h
@@ -1,10 +1,10 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the Apache 2.0 License.
 #pragma once
+#include "ccf/entity_id.h"
 #include "client_signatures.h"
 #include "crypto/pem.h"
 #include "ds/hash.h"
-#include "entity_id.h"
 #include "node_signature.h"
 
 #include <msgpack/msgpack.hpp>
