+++ src/node/nodes.h
@@ -1,8 +1,9 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the Apache 2.0 License.
 #pragma once
+
+#include "ccf/entity_id.h"
 #include "entities.h"
-#include "entity_id.h"
 #include "kv/map.h"
 #include "node_info_network.h"
 #include "quote_info.h"
